rename block to martech-metadata
