--- libs/blocks/martech-attribute-metadata/martech-attribute-metadata.css
@@ -1,1 +1,0 @@
-.martech-attribute-metadata { display: none; }
