+++ libs/blocks/martech-metadata/martech-metadata.css
@@ -1,0 +1,1 @@
+.martech-metadata { display: none; }
